+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.1.6.8
 
+- [api] rewrite of scoped lookup/map
+
 ## 0.1.6.7
 
 ## 0.1.6.6